Fix github action
