--- contrib/sklearn/.gitignore
@@ -1,1 +1,0 @@
-mlruns.db
