+++ contrib/sklearn/setup.py
@@ -13,8 +13,6 @@ setup(
         "numpy",
         "scikit-learn",
     ],
-    extras_require={
-        "dev": ["black", "isort", "pytest", "mlflow", "scikit-learn"]
-    },
+    extras_require={"dev": ["black", "isort", "pytest", "mlflow"]},
     packages=find_namespace_packages(include=["rikai.*"]),
 )
