+++ contrib/sklearn/tests/test_sklearn.py
@@ -12,75 +12,113 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import getpass
+from pathlib import Path
 
 import mlflow
 import numpy as np
-from pyspark.sql import SparkSession
+from pyspark.sql import Row, SparkSession
+from pyspark.sql.types import DoubleType, LongType, StructField, StructType
+from sklearn.datasets import make_classification
+from sklearn.ensemble import RandomForestClassifier
 from sklearn.linear_model import LinearRegression
 
 import rikai
 
 
-def test_linear_regression(spark: SparkSession):
-    mlflow_tracking_uri = "sqlite:///mlruns.db"
-    mlflow.set_tracking_uri(mlflow_tracking_uri)
-
-    # enable autologging
-    mlflow.sklearn.autolog()
-
+def test_sklearn_linear_regression(tmp_path: Path, spark: SparkSession):
     # prepare training data
     X = np.array([[1, 1], [1, 2], [2, 2], [2, 3]])
     y = np.dot(X, np.array([1, 2])) + 3
-
-    # prepare evaluation data
-    X_eval = np.array([[3, 3], [3, 4]])
-    y_eval = np.dot(X_eval, np.array([1, 2])) + 3
-
-    # train a model
     model = LinearRegression()
-    with mlflow.start_run() as run:
-        ####
-        # Part 1: Train the model and register it on MLflow
-        ####
+
+    tmp_path.mkdir(parents=True, exist_ok=True)
+    tracking_uri = "sqlite:///" + str(tmp_path / "tracking.db")
+    mlflow.set_tracking_uri(tracking_uri)
+    with mlflow.start_run():
         model.fit(X, y)
-        metrics = mlflow.sklearn.eval_and_log_metrics(
-            model, X_eval, y_eval, prefix="val_"
-        )
 
-        schema = "float"
-        registered_model_name = f"{getpass.getuser()}_sklearn_lr"
+        reg_model_name = "sklearn_linear_regression"
+        model_name = "sk_lr_m"
         rikai.mlflow.sklearn.log_model(
-            model, "model", schema, registered_model_name=registered_model_name
+            model,
+            artifact_path="model",
+            schema="double",
+            registered_model_name=reg_model_name,
         )
 
-        ####
-        # Part 2: create the model using the registered MLflow uri
-        ####
-        spark.conf.set("spark.sql.execution.arrow.pyspark.enabled", "false")
         spark.conf.set(
-            "rikai.sql.ml.registry.mlflow.tracking_uri", mlflow_tracking_uri
+            "rikai.sql.ml.registry.mlflow.tracking_uri", tracking_uri
         )
         spark.sql(
             f"""
-        CREATE MODEL mlflow_sklearn_m USING 'mlflow:///{registered_model_name}';
-        """
+            CREATE MODEL {model_name} USING 'mlflow:///{reg_model_name}';
+            """
         )
 
-        ####
-        # Part 3: predict using the registered Rikai model
-        ####
-        spark.sql("show models").show(1, vertical=False, truncate=False)
-
-        df = spark.range(100).selectExpr("id as x0", "id+1 as x1")
+        df = spark.range(2).selectExpr("id as x0", "id+1 as x1")
         df.createOrReplaceTempView("tbl_X")
 
         result = spark.sql(
             f"""
-        select ML_PREDICT(mlflow_sklearn_m, array(x0, x1)) from tbl_X
-        """
+            select ML_PREDICT({model_name}, array(x0, x1)) as pred from tbl_X
+            """
+        )
+        assert result.schema == StructType([StructField("pred", DoubleType())])
+        assert result.count() == 2
+
+
+def test_sklearn_random_forest(tmp_path: Path, spark: SparkSession):
+    X, y = make_classification(
+        n_samples=1000,
+        n_features=4,
+        n_informative=2,
+        n_redundant=0,
+        random_state=0,
+        shuffle=False,
+    )
+
+    # train a model
+    model = RandomForestClassifier(max_depth=2, random_state=0)
+
+    tmp_path.mkdir(parents=True, exist_ok=True)
+    tracking_uri = "sqlite:///" + str(tmp_path / "tracking.db")
+    mlflow.set_tracking_uri(tracking_uri)
+    with mlflow.start_run():
+        model.fit(X, y)
+
+        reg_model_name = "sklearn_random_forest"
+        model_name = "sk_rf_m"
+        rikai.mlflow.sklearn.log_model(
+            model,
+            artifact_path="model",
+            schema="long",
+            registered_model_name=reg_model_name,
         )
 
-        result.printSchema()
-        result.show(10, vertical=False, truncate=False)
-        assert result.count() == 100
+        spark.conf.set(
+            "rikai.sql.ml.registry.mlflow.tracking_uri", tracking_uri
+        )
+        spark.sql(
+            f"""
+            CREATE MODEL {model_name} USING 'mlflow:///{reg_model_name}';
+            """
+        )
+
+        df = spark.range(2).selectExpr(
+            "id as x0", "id+1 as x1", "id+2 as x2", "id+3 as x3"
+        )
+        df.createOrReplaceTempView("tbl_X")
+
+        result = spark.sql(
+            f"""
+            select
+                ML_PREDICT({model_name}, array(x0, x1, x2, x3)) as pred
+            from tbl_X
+            """
+        )
+        result.show()
+        assert result.schema == StructType([StructField("pred", LongType())])
+        assert (
+            result.collect()
+            == spark.createDataFrame([Row(pred=1), Row(pred=1)]).collect()
+        )
